refactor: step 6: extract replication, secrets, storage, metrics
Move the remaining topic-specific assertions out of tests/e2e/asserts_test.go
into the four new packages:

- tests/internal/asserts/replication grows to its full size:
  AssertClusterStandbysAreStreaming, AssertStandbysFollowPromotion,
  AssertWritesResumedBeforeTimeout, AssertReplicaModeCluster,
  AssertDetachReplicaModeCluster, AssertFastFailOver,
  AssertReplicationSlotsOnPod, AssertClusterReplicationSlotsAligned,
  AssertClusterHAReplicationSlots.

- tests/internal/asserts/secrets carries AssertUpdateSecret,
  CreateAndAssertServerCertificatesSecrets,
  CreateAndAssertClientCertificatesSecrets and
  AssertSSLVerifyFullDBConnectionFromAppPod.

- tests/internal/asserts/storage carries AssertPvcHasLabels.

- tests/internal/asserts/metrics carries AssertCustomMetricsResourcesExist,
  AssertMetricsData, the two CollectAndAssert*MetricsPresentOnEachPod
  helpers (now exported), and Assert{Includes,Excludes}Metrics.

The four restore variants deferred from step 4
(AssertClusterRestore, AssertClusterRestoreWithApplicationDB,
AssertClusterWasRestoredWithPITRAndApplicationDB, AssertClusterImport)
now also move into tests/internal/asserts/backup, since the streaming-replication
and secrets helpers they depend on are finally available there.

tests/e2e/asserts_test.go is left as a placeholder for the final
removal commit; every helper has now been migrated.
