+++ tests/e2e/certificates_test.go
@@ -32,6 +32,7 @@ import (
 	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
 	"github.com/cloudnative-pg/cloudnative-pg/tests"
 	clusterasserts "github.com/cloudnative-pg/cloudnative-pg/tests/internal/asserts/cluster"
+	secretsasserts "github.com/cloudnative-pg/cloudnative-pg/tests/internal/asserts/secrets"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/clusterutils"
 	podutils "github.com/cloudnative-pg/cloudnative-pg/tests/utils/pods"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/run"
@@ -214,18 +215,17 @@ var _ = Describe("Certificates", func() {
 					defaultCASecretName, kubectlCNPGClientCertSecretName)
 				err := podutils.CreateAndWaitForReady(env.Ctx, env.Client, &pod, 240)
 				Expect(err).ToNot(HaveOccurred())
-				AssertSSLVerifyFullDBConnectionFromAppPod(namespace, clusterName, pod)
+				secretsasserts.AssertSSLVerifyFullDBConnectionFromAppPod(env, namespace, clusterName, pod)
 			})
 
 		It("can authenticate after switching to user-supplied server certs", Label(tests.LabelServiceConnectivity),
 			func() {
-				CreateAndAssertServerCertificatesSecrets(
+				secretsasserts.CreateAndAssertServerCertificatesSecrets(env,
 					namespace,
 					clusterName,
 					serverCASecretName,
 					serverCertSecretName,
-					false,
-				)
+					false)
 
 				var err error
 				// Updating defaults certificates entries with user provided certificates,
@@ -262,13 +262,13 @@ var _ = Describe("Certificates", func() {
 				)
 				err = podutils.CreateAndWaitForReady(env.Ctx, env.Client, &pod, 240)
 				Expect(err).ToNot(HaveOccurred())
-				AssertSSLVerifyFullDBConnectionFromAppPod(namespace, clusterName, pod)
+				secretsasserts.AssertSSLVerifyFullDBConnectionFromAppPod(env, namespace, clusterName, pod)
 			})
 
 		It("can connect after switching to user-supplied client certificates", Label(tests.LabelServiceConnectivity),
 			func() {
 				// Create certificates secret for client
-				CreateAndAssertClientCertificatesSecrets(namespace, clusterName, clientCASecretName,
+				secretsasserts.CreateAndAssertClientCertificatesSecrets(env, namespace, clusterName, clientCASecretName,
 					replicaCertSecretName,
 					clientCertSecretName, false)
 
@@ -296,7 +296,7 @@ var _ = Describe("Certificates", func() {
 				pod := defaultPodFunc(namespace, "app-pod-cert-3", defaultCASecretName, clientCertSecretName)
 				err := podutils.CreateAndWaitForReady(env.Ctx, env.Client, &pod, 240)
 				Expect(err).ToNot(HaveOccurred())
-				AssertSSLVerifyFullDBConnectionFromAppPod(namespace, clusterName, pod)
+				secretsasserts.AssertSSLVerifyFullDBConnectionFromAppPod(env, namespace, clusterName, pod)
 			})
 
 		It("can connect after switching both server and client certificates to user-supplied mode",
@@ -334,7 +334,7 @@ var _ = Describe("Certificates", func() {
 				pod := defaultPodFunc(namespace, "app-pod-cert-4", serverCASecretName, clientCertSecretName)
 				err := podutils.CreateAndWaitForReady(env.Ctx, env.Client, &pod, 240)
 				Expect(err).ToNot(HaveOccurred())
-				AssertSSLVerifyFullDBConnectionFromAppPod(namespace, clusterName, pod)
+				secretsasserts.AssertSSLVerifyFullDBConnectionFromAppPod(env, namespace, clusterName, pod)
 			})
 	})
 
@@ -353,13 +353,13 @@ var _ = Describe("Certificates", func() {
 			// Create a cluster in a namespace that will be deleted after the test
 			namespace, err = env.CreateUniqueTestNamespace(env.Ctx, env.Client, namespacePrefix)
 			Expect(err).ToNot(HaveOccurred())
-			CreateAndAssertServerCertificatesSecrets(
+			secretsasserts.CreateAndAssertServerCertificatesSecrets(env,
 				namespace,
 				clusterName,
 				serverCASecretName,
 				serverCertSecretName,
-				false,
-			)
+				false)
+
 			clusterasserts.AssertCreateCluster(env, testTimeouts, namespace, clusterName, sampleFile)
 			cluster, err := clusterutils.Get(env.Ctx, env.Client, namespace, clusterName)
 			Expect(err).ToNot(HaveOccurred())
@@ -380,7 +380,7 @@ var _ = Describe("Certificates", func() {
 			)
 			err = podutils.CreateAndWaitForReady(env.Ctx, env.Client, &pod, 240)
 			Expect(err).ToNot(HaveOccurred())
-			AssertSSLVerifyFullDBConnectionFromAppPod(namespace, clusterName, pod)
+			secretsasserts.AssertSSLVerifyFullDBConnectionFromAppPod(env, namespace, clusterName, pod)
 		})
 	})
 
@@ -401,19 +401,19 @@ var _ = Describe("Certificates", func() {
 				Expect(err).ToNot(HaveOccurred())
 
 				// Create certificates secret for client
-				CreateAndAssertClientCertificatesSecrets(
+				secretsasserts.CreateAndAssertClientCertificatesSecrets(env,
 					namespace,
 					clusterName,
 					clientCASecretName,
 					replicaCertSecretName,
 					clientCertSecretName,
-					false,
-				)
+					false)
+
 				clusterasserts.AssertCreateCluster(env, testTimeouts, namespace, clusterName, sampleFile)
 				pod := defaultPodFunc(namespace, "app-pod-cert-3", defaultCASecretName, clientCertSecretName)
 				err = podutils.CreateAndWaitForReady(env.Ctx, env.Client, &pod, 240)
 				Expect(err).ToNot(HaveOccurred())
-				AssertSSLVerifyFullDBConnectionFromAppPod(namespace, clusterName, pod)
+				secretsasserts.AssertSSLVerifyFullDBConnectionFromAppPod(env, namespace, clusterName, pod)
 			})
 	})
 
@@ -434,27 +434,26 @@ var _ = Describe("Certificates", func() {
 				Expect(err).ToNot(HaveOccurred())
 
 				// Create certificates secret for server
-				CreateAndAssertServerCertificatesSecrets(
+				secretsasserts.CreateAndAssertServerCertificatesSecrets(env,
 					namespace,
 					clusterName,
 					serverCASecretName,
 					serverCertSecretName,
-					false,
-				)
+					false)
 
-				CreateAndAssertClientCertificatesSecrets(
+				secretsasserts.CreateAndAssertClientCertificatesSecrets(env,
 					namespace,
 					clusterName,
 					clientCASecretName,
 					replicaCertSecretName,
 					clientCertSecretName,
-					false,
-				)
+					false)
+
 				clusterasserts.AssertCreateCluster(env, testTimeouts, namespace, clusterName, sampleFile)
 				pod := defaultPodFunc(namespace, "app-pod-cert-4", serverCASecretName, clientCertSecretName)
 				err = podutils.CreateAndWaitForReady(env.Ctx, env.Client, &pod, 240)
 				Expect(err).ToNot(HaveOccurred())
-				AssertSSLVerifyFullDBConnectionFromAppPod(namespace, clusterName, pod)
+				secretsasserts.AssertSSLVerifyFullDBConnectionFromAppPod(env, namespace, clusterName, pod)
 			})
 	})
 })
