+++ tests/e2e/pg_wal_volume_test.go
@@ -34,6 +34,7 @@ import (
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
 	"github.com/cloudnative-pg/cloudnative-pg/tests"
 	clusterasserts "github.com/cloudnative-pg/cloudnative-pg/tests/internal/asserts/cluster"
+	storageasserts "github.com/cloudnative-pg/cloudnative-pg/tests/internal/asserts/storage"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/clusterutils"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/exec"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/timeouts"
@@ -66,7 +67,7 @@ var _ = Describe("Separate pg_wal volume", Label(tests.LabelStorage), func() {
 				err := env.Client.Get(env.Ctx, namespacedPVCName, pvc)
 				Expect(pvc.GetName(), err).To(BeEquivalentTo(pvcName))
 			}
-			AssertPvcHasLabels(namespace, clusterName)
+			storageasserts.AssertPvcHasLabels(env, namespace, clusterName)
 		})
 		By("checking that pg_wal is a symlink to the dedicated volume", func() {
 			for _, pod := range podList.Items {
