+++ tests/e2e/replication_slot_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
 	"github.com/cloudnative-pg/cloudnative-pg/tests"
 	clusterasserts "github.com/cloudnative-pg/cloudnative-pg/tests/internal/asserts/cluster"
+	replicationasserts "github.com/cloudnative-pg/cloudnative-pg/tests/internal/asserts/replication"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/clusterutils"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/exec"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/postgres"
@@ -84,7 +85,7 @@ var _ = Describe("Replication Slot", Label(tests.LabelReplication), func() {
 				primaryPod.GetName(),
 			)
 			Expect(err).ToNot(HaveOccurred())
-			AssertReplicationSlotsOnPod(namespace, clusterName, *primaryPod, expectedSlots, true, false)
+			replicationasserts.AssertReplicationSlotsOnPod(env, namespace, clusterName, *primaryPod, expectedSlots, true, false)
 		})
 
 		By("checking standbys HA slots exist", func() {
@@ -100,13 +101,13 @@ var _ = Describe("Replication Slot", Label(tests.LabelReplication), func() {
 					namespace, clusterName, pod.GetName(),
 				)
 				Expect(err).ToNot(HaveOccurred())
-				AssertReplicationSlotsOnPod(namespace, clusterName, pod, expectedSlots, true, false)
+				replicationasserts.AssertReplicationSlotsOnPod(env, namespace, clusterName, pod, expectedSlots, true, false)
 			}
 			GinkgoWriter.Println("Replica pods slot check succeeded in", time.Since(before))
 		})
 
 		By("checking all the slots restart_lsn's are aligned", func() {
-			AssertClusterReplicationSlotsAligned(namespace, clusterName)
+			replicationasserts.AssertClusterReplicationSlotsAligned(env, namespace, clusterName)
 		})
 
 		By("creating a physical replication slots on the primary", func() {
@@ -133,7 +134,15 @@ var _ = Describe("Replication Slot", Label(tests.LabelReplication), func() {
 			}, 90, 2).Should(Succeed())
 			before := time.Now()
 			for _, pod := range replicaPods.Items {
-				AssertReplicationSlotsOnPod(namespace, clusterName, pod, []string{userPhysicalSlot}, false, false)
+				replicationasserts.AssertReplicationSlotsOnPod(
+					env,
+					namespace,
+					clusterName,
+					pod,
+					[]string{userPhysicalSlot},
+					false,
+					false,
+				)
 			}
 			GinkgoWriter.Println("Slot:", userPhysicalSlot, "synchronized to replica pods in", time.Since(before))
 		})
