+++ tests/e2e/update_user_test.go
@@ -32,6 +32,7 @@ import (
 	"github.com/cloudnative-pg/cloudnative-pg/tests"
 	clusterasserts "github.com/cloudnative-pg/cloudnative-pg/tests/internal/asserts/cluster"
 	pgasserts "github.com/cloudnative-pg/cloudnative-pg/tests/internal/asserts/postgres"
+	secretsasserts "github.com/cloudnative-pg/cloudnative-pg/tests/internal/asserts/secrets"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/clusterutils"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/exec"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/postgres"
@@ -77,7 +78,7 @@ var _ = Describe("Update user and superuser password", Label(tests.LabelServiceC
 		By("update user application password", func() {
 			const newPassword = "eeh2Zahohx"
 
-			AssertUpdateSecret("password", newPassword, appSecretName, namespace, clusterName, 30, env)
+			secretsasserts.AssertUpdateSecret(env, "password", newPassword, appSecretName, namespace, clusterName, 30)
 			pgasserts.AssertConnection(env, namespace, rwService, postgres.AppDBName, postgres.AppUser, newPassword)
 		})
 
@@ -85,8 +86,8 @@ var _ = Describe("Update user and superuser password", Label(tests.LabelServiceC
 			const newUser = "postgres"
 			const newPassword = "newpassword"
 
-			AssertUpdateSecret("password", newPassword, appSecretName, namespace, clusterName, 30, env)
-			AssertUpdateSecret("username", newUser, appSecretName, namespace, clusterName, 30, env)
+			secretsasserts.AssertUpdateSecret(env, "password", newPassword, appSecretName, namespace, clusterName, 30)
+			secretsasserts.AssertUpdateSecret(env, "username", newUser, appSecretName, namespace, clusterName, 30)
 
 			timeout := time.Second * 10
 			dsn := services.CreateDSN(rwService, newUser, postgres.AppDBName, newPassword, services.Require, 5432)
@@ -97,7 +98,7 @@ var _ = Describe("Update user and superuser password", Label(tests.LabelServiceC
 			Expect(err).To(HaveOccurred())
 
 			// Revert the username change
-			AssertUpdateSecret("username", postgres.AppUser, appSecretName, namespace, clusterName, 30, env)
+			secretsasserts.AssertUpdateSecret(env, "username", postgres.AppUser, appSecretName, namespace, clusterName, 30)
 		})
 
 		By("update superuser password", func() {
@@ -121,7 +122,7 @@ var _ = Describe("Update user and superuser password", Label(tests.LabelServiceC
 			}, 60).Should(Succeed())
 
 			const newPassword = "fi6uCae7"
-			AssertUpdateSecret("password", newPassword, superUserSecretName, namespace, clusterName, 30, env)
+			secretsasserts.AssertUpdateSecret(env, "password", newPassword, superUserSecretName, namespace, clusterName, 30)
 			pgasserts.AssertConnection(env, namespace, rwService, postgres.PostgresDBName, postgres.PostgresUser, newPassword)
 		})
 	})
